02_file_handling_and_dataframes/02_challenge/data_transform.py
import csv
import json
import os

def read_csv(file_path):
    """
    Reads a CSV file and returns its contents as a list of dictionaries.

    :param file_path: str - The path to the CSV file.
    :return: list - A list of dictionaries representing the CSV rows.
    """
    # Implementation goes here
    pass

def csv_to_json(csv_data):
    """
    Takes CSV data (list of dictionaries) and converts it to JSON format (string).

    :param csv_data: list - The CSV data as a list of dictionaries.
    :return: str - The JSON representation of the data.
    """
    # Implementation goes here
    pass

def write_json(json_data, file_path):
    """
    Writes JSON data to a file.

    :param json_data: str - The JSON data to write.
    :param file_path: str - The path to the JSON file.
    """
    # Implementation goes here
    pass

def read_json(file_path):
    """
    Reads a JSON file and returns its contents.

    :param file_path: str - The path to the JSON file.
    :return: The JSON file's contents.
    """
    # Implementation goes here
    pass

def json_to_csv(json_data):
    """
    Takes JSON data (typically a list of dictionaries) and converts it to CSV format (string).

    :param json_data: The JSON data.
    :return: str - The CSV representation of the data.
    """
    # Implementation goes here
    pass

def write_csv(csv_data, file_path):
    """
    Writes CSV data to a file.

    :param csv_data: str - The CSV data to write.
    :param file_path: str - The path to the CSV file.
    """
    # Implementation goes here
    pass

def validate_data(data, data_type):
    """
    Checks data integrity (e.g., consistent number of columns in CSV).

    :param data: The data to validate.
    :param data_type: str - The type of data ('CSV' or 'JSON').
    :return: bool - True if data is valid, False otherwise.
    """
    # Implementation goes here
    pass

def process_directory(directory_path):
    """
    Identifies all CSV and JSON files in a given directory and performs appropriate conversions.

    :param directory_path: str - The path to the directory to process.
    """
    # Implementation goes here
    pass

# Main function to run the script
def main():
    # Example usage
    try:
        directory = "path_to_directory"
        process_directory(directory)
    except Exception as e:
        print("An error occurred:", e)

if __name__ == "__main__":
    main()