02_file_handling_and_dataframes/02_challenge/test_data_transform.py
import unittest
import os
from data_transform import read_csv, csv_to_json, write_json, read_json, json_to_csv, write_csv, validate_data, process_directory

class TestReadCSV(unittest.TestCase):
    def test_read_valid_csv(self):
        result = read_csv('sample.csv')
        self.assertIsInstance(result, list)
        self.assertGreater(len(result), 0)

    def test_read_nonexistent_csv(self):
        with self.assertRaises(FileNotFoundError):
            read_csv('nonexistent.csv')

# Similar test classes for csv_to_json, write_json, read_json, json_to_csv, write_csv, validate_data, process_directory

class TestCsvToJson(unittest.TestCase):
    def test_valid_csv_to_json(self):
        csv_data = [{'name': 'John', 'age': 30}, {'name': 'Jane', 'age': 25}]
        result = csv_to_json(csv_data)
        self.assertIsInstance(result, str)

class TestWriteJson(unittest.TestCase):
    def test_write_json(self):
        json_data = '[{"name": "John", "age": 30}, {"name": "Jane", "age": 25}]'
        file_path = 'test_output.json'
        write_json(json_data, file_path)
        self.assertTrue(os.path.exists(file_path))
        os.remove(file_path)

class TestReadJson(unittest.TestCase):
    def test_read_valid_json(self):
        result = read_json('sample.json')
        self.assertIsInstance(result, list)

class TestJsonToCsv(unittest.TestCase):
    def test_valid_json_to_csv(self):
        json_data = [{"name": "John", "age": 30}, {"name": "Jane", "age": 25}]
        result = json_to_csv(json_data)
        self.assertIsInstance(result, str)

class TestWriteCsv(unittest.TestCase):
    def test_write_csv(self):
        csv_data = 'name,age\nJohn,30\nJane,25'
        file_path = 'test_output.csv'
        write_csv(csv_data, file_path)
        self.assertTrue(os.path.exists(file_path))
        os.remove(file_path)

class TestValidateData(unittest.TestCase):
    def test_validate_valid_csv_data(self):
        csv_data = [{'name': 'John', 'age': 30}, {'name': 'Jane', 'age': 25}]
        result = validate_data(csv_data, 'CSV')
        self.assertTrue(result)

class TestProcessDirectory(unittest.TestCase):
    # This test depends on the implementation of process_directory
    pass

if __name__ == '__main__':
    unittest.main()
